:bulb: chore: using react-loadable
