+++ .eslintrc.js
@@ -8,12 +8,14 @@ module.exports = {
     Atomics: 'readonly',
     SharedArrayBuffer: 'readonly',
   },
+  parser: 'babel-eslint',
   parserOptions: {
     ecmaFeatures: {
       jsx: true,
     },
     ecmaVersion: 2018,
     sourceType: 'module',
+    allowImportExportEverywhere: true,
   },
   plugins: [
     'react',
