+++ gatsby-config.js
@@ -58,6 +58,6 @@ module.exports = {
         '~': path.join(__dirname, 'src'),
       },
     },
-    // `gatsby-plugin-offline`,
+    `gatsby-plugin-offline`,
   ],
 };
