+++ package.json
@@ -30,10 +30,12 @@
     "react-dom": "^16.11.0",
     "react-helmet": "^5.2.1",
     "react-icons": "^3.8.0",
+    "react-loadable": "^5.5.0",
     "styled-components": "^4.4.1",
     "typeface-raleway": "^0.0.75"
   },
   "devDependencies": {
+    "babel-eslint": "^10.0.3",
     "cross-env": "^6.0.3",
     "eslint": "^6.6.0",
     "eslint-config-airbnb": "^18.0.1",
