+++ src/atoms/Loading/index.js
@@ -1,0 +1,10 @@
+import React from 'react';
+// import { FiLoader } from 'react-icons/fi';
+
+import { Container } from './styles';
+
+function Loading() {
+  return <Container>{/* <FiLoader size={80} /> */}</Container>;
+}
+
+export default Loading;
