+++ src/organismis/About/index.js
@@ -37,8 +37,9 @@ function About({ forwardRef }) {
 }
 
 About.propTypes = {
-  forwardRef: PropTypes.shape({ current: PropTypes.instanceOf(Element) })
-    .isRequired,
+  forwardRef: PropTypes.shape({
+    current: PropTypes.object,
+  }).isRequired,
 };
 
 export default About;
