+++ src/organismis/Education/index.js
@@ -18,8 +18,9 @@ function Education({ forwardRef }) {
 }
 
 Education.propTypes = {
-  forwardRef: PropTypes.shape({ current: PropTypes.instanceOf(Element) })
-    .isRequired,
+  forwardRef: PropTypes.shape({
+    current: PropTypes.object,
+  }).isRequired,
 };
 
 export default Education;
