+++ src/pages/404/index.js
@@ -1,24 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 
 import { Container, Button } from './styles';
 
 function NotFound() {
-  const [height, setHeight] = useState(window.innerHeight);
-
-  useEffect(() => {
-    function setCurrentHeight() {
-      setHeight(window.innerHeight);
-    }
-
-    window.addEventListener('resize', setCurrentHeight);
-
-    return () => {
-      window.removeEventListener('resize', setCurrentHeight);
-    };
-  }, [height]);
-
   return (
-    <Container title="Not found" height={height}>
+    <Container title="Not found">
       <h3>Oops!!</h3>
       <main>
         <h1>
